update suite page to match new tests ui
