+++ app/components/Dashboard/Suite/List/index.tsx
@@ -1,7 +1,6 @@
 import { Box } from "grommet";
 
-import { SuiteRun } from "../../../../lib/types";
-import { borderSize } from "../../../../theme/theme";
+import { Suite, SuiteRun } from "../../../../lib/types";
 import Header from "./Header";
 import RunCard from "./RunCard";
 
@@ -9,12 +8,14 @@ type Props = {
   checkedTestIds: string[];
   runs: SuiteRun[];
   setCheckedTestIds: (runIds: string[]) => void;
+  suite: Suite;
 };
 
 export default function List({
   checkedTestIds,
   runs,
   setCheckedTestIds,
+  suite,
 }: Props): JSX.Element {
   const handleRunCheck = (testId: string): void => {
     const index = checkedTestIds.indexOf(testId);
@@ -28,12 +29,11 @@ export default function List({
     }
   };
 
-  const runsHtml = runs.map((run, i) => {
+  const runsHtml = runs.map((run) => {
     return (
       <RunCard
         isChecked={checkedTestIds.includes(run.test_id)}
         key={run.id}
-        noBorder={!i}
         onCheck={() => handleRunCheck(run.test_id)}
         run={run}
       />
@@ -41,19 +41,16 @@ export default function List({
   });
 
   return (
-    <Box
-      border={{ color: "gray3", size: borderSize.xsmall }}
-      margin={{ top: "medium" }}
-      round={borderSize.small}
-    >
+    <>
       <Header
         checkedTestIds={checkedTestIds}
         runs={runs}
         setCheckedTestIds={setCheckedTestIds}
+        suite={suite}
       />
       <Box overflow={{ vertical: "auto" }}>
         <Box flex={false}>{runsHtml}</Box>
       </Box>
-    </Box>
+    </>
   );
 }
